Move FakeDaprActorClient to fake_client module
