+++ ext/flask_dapr/setup.cfg
@@ -11,6 +11,7 @@ classifiers =
     Operating System :: OS Independent
     Programming Language :: Python
     Programming Language :: Python :: 3.7
+    Programming Language :: Python :: 3.8
 project_urls =
     Documentation = https://github.com/dapr/docs
     Source = https://github.com/dapr/python-sdk
