+++ tests/actor/test_state_manager.py
@@ -18,8 +18,7 @@ from dapr.actor.runtime._type_information import ActorTypeInformation
 from dapr.serializers import DefaultJSONSerializer
 
 from tests.actor.fake_actor_classes import FakeSimpleActor
-
-from tests.actor.fake_actor_classes import FakeDaprActorClient
+from tests.actor.fake_client import FakeDaprActorClient
 
 from tests.actor.utils import (
     _async_mock,
@@ -40,10 +39,10 @@ class ActorStateManagerTests(unittest.TestCase):
         self._fake_actor = FakeSimpleActor(self._runtime_ctx, self._test_actor_id)
 
     @mock.patch(
-        'tests.actor.fake_actor_classes.FakeDaprActorClient.get_state',
+        'tests.actor.fake_client.FakeDaprActorClient.get_state',
         new=_async_mock(return_value=base64.b64encode(b'"value1"')))
     @mock.patch(
-        'tests.actor.fake_actor_classes.FakeDaprActorClient.save_state_transactionally',
+        'tests.actor.fake_client.FakeDaprActorClient.save_state_transactionally',
         new=_async_mock())
     def test_add_state(self):
 
@@ -61,7 +60,7 @@ class ActorStateManagerTests(unittest.TestCase):
         added = _run(state_manager.try_add_state('state1', 'value1'))
         self.assertFalse(added)
 
-    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=_async_mock())
+    @mock.patch('tests.actor.fake_client.FakeDaprActorClient.get_state', new=_async_mock())
     def test_get_state_for_no_state(self):
 
         state_manager = ActorStateManager(self._fake_actor)
@@ -76,7 +75,7 @@ class ActorStateManagerTests(unittest.TestCase):
         self.assertIsNone(val)
 
     @mock.patch(
-        'tests.actor.fake_actor_classes.FakeDaprActorClient.get_state',
+        'tests.actor.fake_client.FakeDaprActorClient.get_state',
         new=_async_mock(return_value=b'"value1"'))
     def test_get_state_for_existing_value(self):
 
@@ -86,7 +85,7 @@ class ActorStateManagerTests(unittest.TestCase):
         self.assertEqual("value1", val)
 
     @mock.patch(
-        'tests.actor.fake_actor_classes.FakeDaprActorClient.get_state',
+        'tests.actor.fake_client.FakeDaprActorClient.get_state',
         new=_async_mock(return_value=b'"value1"'))
     def test_get_state_for_removed_value(self):
 
@@ -101,7 +100,7 @@ class ActorStateManagerTests(unittest.TestCase):
         self.assertFalse(has_value)
         self.assertIsNone(val)
 
-    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=_async_mock())
+    @mock.patch('tests.actor.fake_client.FakeDaprActorClient.get_state', new=_async_mock())
     def test_set_state_for_new_state(self):
 
         state_manager = ActorStateManager(self._fake_actor)
@@ -111,7 +110,7 @@ class ActorStateManagerTests(unittest.TestCase):
         self.assertEqual(StateChangeKind.add, state.change_kind)
         self.assertEqual('value1', state.value)
 
-    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=_async_mock())
+    @mock.patch('tests.actor.fake_client.FakeDaprActorClient.get_state', new=_async_mock())
     def test_set_state_for_existing_state_only_in_mem(self):
 
         state_manager = ActorStateManager(self._fake_actor)
@@ -127,7 +126,7 @@ class ActorStateManagerTests(unittest.TestCase):
         self.assertEqual('value2', state.value)
 
     @mock.patch(
-        'tests.actor.fake_actor_classes.FakeDaprActorClient.get_state',
+        'tests.actor.fake_client.FakeDaprActorClient.get_state',
         new=_async_mock(return_value=b'"value1"'))
     def test_set_state_for_existing_state(self):
 
@@ -137,7 +136,7 @@ class ActorStateManagerTests(unittest.TestCase):
         self.assertEqual(StateChangeKind.update, state.change_kind)
         self.assertEqual('value2', state.value)
 
-    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=_async_mock())
+    @mock.patch('tests.actor.fake_client.FakeDaprActorClient.get_state', new=_async_mock())
     def test_remove_state_for_non_existing_state(self):
 
         state_manager = ActorStateManager(self._fake_actor)
@@ -145,7 +144,7 @@ class ActorStateManagerTests(unittest.TestCase):
         self.assertFalse(removed)
 
     @mock.patch(
-        'tests.actor.fake_actor_classes.FakeDaprActorClient.get_state',
+        'tests.actor.fake_client.FakeDaprActorClient.get_state',
         new=_async_mock(return_value=b'"value1"'))
     def test_remove_state_for_existing_state(self):
 
@@ -156,7 +155,7 @@ class ActorStateManagerTests(unittest.TestCase):
         state = state_manager._state_change_tracker['state1']
         self.assertEqual(StateChangeKind.remove, state.change_kind)
 
-    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=_async_mock())
+    @mock.patch('tests.actor.fake_client.FakeDaprActorClient.get_state', new=_async_mock())
     def test_remove_state_for_existing_state_in_mem(self):
 
         state_manager = ActorStateManager(self._fake_actor)
@@ -164,7 +163,7 @@ class ActorStateManagerTests(unittest.TestCase):
         removed = _run(state_manager.try_remove_state('state1'))
         self.assertTrue(removed)
 
-    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=_async_mock())
+    @mock.patch('tests.actor.fake_client.FakeDaprActorClient.get_state', new=_async_mock())
     def test_remove_state_twice_for_existing_state_in_mem(self):
 
         state_manager = ActorStateManager(self._fake_actor)
@@ -174,7 +173,7 @@ class ActorStateManagerTests(unittest.TestCase):
         removed = _run(state_manager.try_remove_state('state1'))
         self.assertFalse(removed)
 
-    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=_async_mock())
+    @mock.patch('tests.actor.fake_client.FakeDaprActorClient.get_state', new=_async_mock())
     def test_contains_state_for_removed_state(self):
 
         state_manager = ActorStateManager(self._fake_actor)
@@ -184,7 +183,7 @@ class ActorStateManagerTests(unittest.TestCase):
         self.assertTrue(exist)
 
     @mock.patch(
-        'tests.actor.fake_actor_classes.FakeDaprActorClient.get_state',
+        'tests.actor.fake_client.FakeDaprActorClient.get_state',
         new=_async_mock(return_value=b'"value1"'))
     def test_contains_state_for_existing_state(self):
 
@@ -193,7 +192,7 @@ class ActorStateManagerTests(unittest.TestCase):
         self.assertTrue(exist)
 
     @mock.patch(
-        'tests.actor.fake_actor_classes.FakeDaprActorClient.get_state',
+        'tests.actor.fake_client.FakeDaprActorClient.get_state',
         new=_async_mock(return_value=b'"value1"'))
     def test_get_or_add_state_for_existing_state(self):
 
@@ -202,7 +201,7 @@ class ActorStateManagerTests(unittest.TestCase):
         self.assertEqual('value1', val)
 
     @mock.patch(
-        'tests.actor.fake_actor_classes.FakeDaprActorClient.get_state',
+        'tests.actor.fake_client.FakeDaprActorClient.get_state',
         new=_async_mock())
     def test_get_or_add_state_for_non_existing_state(self):
         state_manager = ActorStateManager(self._fake_actor)
@@ -217,7 +216,7 @@ class ActorStateManagerTests(unittest.TestCase):
             self._test_actor_id.id, 'state1')
 
     @mock.patch(
-        'tests.actor.fake_actor_classes.FakeDaprActorClient.get_state',
+        'tests.actor.fake_client.FakeDaprActorClient.get_state',
         new=_async_mock(return_value=b'"value1"'))
     def test_get_or_add_state_for_removed_state(self):
 
@@ -231,7 +230,7 @@ class ActorStateManagerTests(unittest.TestCase):
         self.assertEqual(StateChangeKind.update, state.change_kind)
         self.assertEqual('value2', val)
 
-    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=_async_mock())
+    @mock.patch('tests.actor.fake_client.FakeDaprActorClient.get_state', new=_async_mock())
     def test_add_or_update_state_for_new_state(self):
         """adds state if state does not exist."""
         def test_update_value(name, value):
@@ -244,7 +243,7 @@ class ActorStateManagerTests(unittest.TestCase):
         self.assertEqual(StateChangeKind.add, state.change_kind)
 
     @mock.patch(
-        'tests.actor.fake_actor_classes.FakeDaprActorClient.get_state',
+        'tests.actor.fake_client.FakeDaprActorClient.get_state',
         new=_async_mock(return_value=b'"value1"'))
     def test_add_or_update_state_for_state_in_storage(self):
         """updates state value using update_value_factory if state is
@@ -259,7 +258,7 @@ class ActorStateManagerTests(unittest.TestCase):
         self.assertEqual(StateChangeKind.update, state.change_kind)
 
     @mock.patch(
-        'tests.actor.fake_actor_classes.FakeDaprActorClient.get_state',
+        'tests.actor.fake_client.FakeDaprActorClient.get_state',
         new=_async_mock(return_value=b'"value1"'))
     def test_add_or_update_state_for_removed_state(self):
         """add state value if state was removed."""
@@ -273,7 +272,7 @@ class ActorStateManagerTests(unittest.TestCase):
         self.assertEqual('value1', val)
 
     @mock.patch(
-        'tests.actor.fake_actor_classes.FakeDaprActorClient.get_state',
+        'tests.actor.fake_client.FakeDaprActorClient.get_state',
         new=_async_mock(return_value=b'"value1"'))
     def test_add_or_update_state_for_none_state_key(self):
         """update state value for StateChangeKind.none state """
@@ -294,7 +293,7 @@ class ActorStateManagerTests(unittest.TestCase):
         with self.assertRaises(AttributeError):
             _run(state_manager.add_or_update_state('state1', 'value1', None))
 
-    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=_async_mock())
+    @mock.patch('tests.actor.fake_client.FakeDaprActorClient.get_state', new=_async_mock())
     def test_get_state_names(self):
         state_manager = ActorStateManager(self._fake_actor)
         _run(state_manager.set_state('state1', 'value1'))
@@ -317,7 +316,7 @@ class ActorStateManagerTests(unittest.TestCase):
             'state3')
 
     @mock.patch(
-        'tests.actor.fake_actor_classes.FakeDaprActorClient.get_state',
+        'tests.actor.fake_client.FakeDaprActorClient.get_state',
         new=_async_mock(return_value=b'"value0"'))
     def test_clear_cache(self):
         state_manager = ActorStateManager(self._fake_actor)
@@ -329,10 +328,10 @@ class ActorStateManagerTests(unittest.TestCase):
         self.assertEqual(0, len(state_manager._state_change_tracker))
 
     @mock.patch(
-        'tests.actor.fake_actor_classes.FakeDaprActorClient.get_state',
+        'tests.actor.fake_client.FakeDaprActorClient.get_state',
         new=_async_mock(return_value=b'"value3"'))
     @mock.patch(
-        'tests.actor.fake_actor_classes.FakeDaprActorClient.save_state_transactionally',
+        'tests.actor.fake_client.FakeDaprActorClient.save_state_transactionally',
         new=_async_mock())
     def test_save_state(self):
         state_manager = ActorStateManager(self._fake_actor)
